requirements.txt
asgiref==3.5.2
Django==4.0.6
django-crispy-forms==1.14.0
django-regex-field==2.0.2
Pillow==9.2.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
xmltodict==0.13.0
